Fix notification button issues
