+++ src/components/Notifications/NotificationButton.scss
@@ -1,0 +1,55 @@
+.notification-group {
+    position: relative;
+    margin-left: auto;
+    margin-right: 0px;
+}
+
+.notification-icon {
+    width: 32px;
+    height: auto;
+}
+
+.notification-popup {
+    background-color: rgba(149, 96, 168, 0.7);
+    position: absolute;
+    right: 50px;
+    top: 0px;
+
+    text-align: center;
+    padding-left: 50px;
+    padding-right: 50px;
+    padding-top: 15px;
+    padding-bottom: 15px;
+    border-radius: 30px;
+
+    width: 400px;
+
+    font-family: "Montserrat";
+    color: white;
+
+    z-index: 4;
+
+    .notification-text {
+        font-family: Montserrat;
+        font-size: 18px;
+        font-weight: 600;
+        margin-bottom: 5px;
+    }
+
+    .notification-action-btn {
+        font-size: 12px;
+        width: 120px;
+        height: 30px;
+        border-radius: 20px;
+    }
+
+    .notification-close-btn {
+        position: absolute;
+        right: 10px;
+        top: 5px;
+
+        padding: 2px 5px 2px 5px;
+        background: transparent;
+        font-weight: 600;
+    }
+}
