+++ src/components/Notifications/ViewNotifications.scss
@@ -1,3 +1,7 @@
+.notification-btn-holder {
+    display: flex;
+}
+
 .notification-header {
     padding-top: 25px;
